feat: stash code
